pipeline: Phase 4 — migrate bossgremlin into pipeline.orchestrators.boss
Port skills/bossgremlin/bossgremlin.py into pipeline/orchestrators/boss.py,
preserving the boss_state.json schema byte-for-byte so in-flight chains work
across the migration. Wire `python -m pipeline.cli boss` in cli.py, update
_bg/launch.sh (both fresh-launch and --resume dispatch paths) to route
bossgremlin through the pipeline, and replace the old skill script with a
thin shim. Add 23 tests in test_orchestrator_boss.py covering child
sequencing, handoff signal parsing, operator-task filtering, resume path
(with a sanitized live fixture), and rescue-then-land.

Closes #118
