+++ pipeline/tests/fixtures/boss_state_sample.json
@@ -1,0 +1,99 @@
+{
+  "spec_path": "/tmp/issue80-spec.md",
+  "chain_kind": "gh",
+  "chain_base_ref": "b06a4ad98aace7ce62740cbe2dbb50c89ce41a73",
+  "target_branch": "main",
+  "current_plan": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-005.md",
+  "handoff_count": 5,
+  "current_child_id": "migrate-bossgremlin-into-pipeline-48ab50",
+  "children": [
+    {
+      "id": "phase-0-inventory-gremlin-sources-and-3b9c15",
+      "outcome": "landed"
+    },
+    {
+      "id": "scaffold-pipeline-package-and-migrate-0b2870",
+      "outcome": "landed"
+    },
+    {
+      "id": "pipeline-phase-2-tests-against-the-new-4d3130",
+      "outcome": "rescued-then-landed"
+    },
+    {
+      "id": "pipeline-phase-3-migrate-ghgremlin-into-f23662",
+      "outcome": "rescued-then-landed"
+    }
+  ],
+  "handoff_records": [
+    {
+      "timestamp": "2026-04-26T19:54:57Z",
+      "n": 1,
+      "plan_in": "/tmp/issue80-spec.md",
+      "plan_out": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-001.md",
+      "signal_file": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-001.state.json",
+      "exit_state": "next-plan",
+      "child_plan": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-001-child.md",
+      "bail_reason": null,
+      "operator_followups": [
+        "After Phase 1 lands: run scripts/sync.sh push --dry-run to confirm pipeline/ doesn't trip DELETE_THRESHOLD"
+      ]
+    },
+    {
+      "timestamp": "2026-04-26T21:08:13Z",
+      "n": 2,
+      "plan_in": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-001.md",
+      "plan_out": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-002.md",
+      "signal_file": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-002.state.json",
+      "exit_state": "next-plan",
+      "child_plan": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-002-child.md",
+      "bail_reason": null,
+      "operator_followups": [
+        "After Phase 1 lands: Run scripts/sync.sh push --dry-run to confirm the new pipeline/ dir doesn't trip DELETE_THRESHOLD"
+      ]
+    },
+    {
+      "timestamp": "2026-04-26T21:37:38Z",
+      "n": 3,
+      "plan_in": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-002.md",
+      "plan_out": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-003.md",
+      "signal_file": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-003.state.json",
+      "exit_state": "next-plan",
+      "child_plan": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-003-child.md",
+      "bail_reason": null,
+      "operator_followups": [
+        "After Phase 1 lands: Run scripts/sync.sh push --dry-run to confirm the new pipeline/ dir doesn't trip DELETE_THRESHOLD"
+      ]
+    },
+    {
+      "timestamp": "2026-04-26T22:23:55Z",
+      "n": 4,
+      "plan_in": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-003.md",
+      "plan_out": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-004.md",
+      "signal_file": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-004.state.json",
+      "exit_state": "next-plan",
+      "child_plan": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-004-child.md",
+      "bail_reason": null,
+      "operator_followups": [
+        "After Phase 1 lands (pending): Run scripts/sync.sh push --dry-run"
+      ]
+    },
+    {
+      "timestamp": "2026-04-27T00:20:20Z",
+      "n": 5,
+      "plan_in": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-004.md",
+      "plan_out": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-005.md",
+      "signal_file": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-005.state.json",
+      "exit_state": "next-plan",
+      "child_plan": "/tmp/claude-gremlins/extract-pipeline-package-via-phased-d04390/handoff-005-child.md",
+      "bail_reason": null,
+      "operator_followups": [
+        "After Phase 4 lands: Confirm /gremlins rescue <boss-id> works against an in-flight chain",
+        "After Phase 4 lands: Run scripts/e2e-bossgremlin.sh against the migrated boss"
+      ]
+    }
+  ],
+  "operator_followups": [
+    "After Phase 4 lands: Confirm /gremlins rescue <boss-id> works against an in-flight chain",
+    "After Phase 4 lands: Run scripts/e2e-bossgremlin.sh against the migrated boss"
+  ]
+}
